refactor api clients into separate modules
